pyproject.toml
[tool.poetry]
name = "doby"
version = "0.1.0"
description = "Doby"
authors = ["Your Name <you@example.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.28.0"
django = "5.2"
django-rest-framework = {extras = ["sidecar"], version = "^0.1.0"}
rest-framework-simplejwt = "^0.0.2"
djangorestframework-simplejwt = "^5.5.1"
channels = "^4.3.2"
psycopg2 = "^2.9.11"
psycopg2-binary = "^2.9.11"
pillow = "^12.1.1"
drf-spectacular = "^0.29.0"
sidecar = "^0.8.1"
drf-spectacular-sidecar = "^2026.3.1"
daphne = "^4.2.1"
channels-redis = "^4.3.0"
celery = "^5.6.3"
redis = "^7.4.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
black = "^23.0.0"
pytest-asyncio = "^1.3.0"
pytest-django = "^4.12.0"
factory-boy = "^3.3.3"
pytest-factoryboy = "^2.8.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
